backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
flake8==6.0.0
python-dotenv==1.0.1
Pillow==9.0.0
django-debug-toolbar==3.2.3